requirements.txt
streamlit
python-docx
lxml
pillow
